new: basics about props
