+++ src/components/HelloWorld.js
@@ -1,0 +1,10 @@
+function HelloWorld() {
+
+    return (
+        <div>
+            <h1>my component</h1>
+        </div>
+    )
+}
+
+export default HelloWorld;
