+++ src/components/SayMyName.js
@@ -1,0 +1,10 @@
+function SayMyName(props) {
+
+    return(
+        <div>
+            <p>Fala aí, {props.nome}! Suave?</p>
+        </div>
+    )
+}
+
+export default SayMyName;
